changing data type
git-svn-id: https://svn.jnitinc.com/svn/vsdev/branches/VitalSignsMongo@918 6b605366-3ce6-ac4f-954a-f758cfb76b34

+++ VSNext.Mongo/VSNext.Mongo.Entities/Status.cs
@@ -1404,7 +1404,7 @@ namespace VSNext.Mongo.Entities
         [DataMember]
         [BsonIgnoreIfNull]
         [BsonElement("threshold")]
-        public int? Threshold { get; set; }
+        public double? Threshold { get; set; }
 
         [DataMember]
         [BsonIgnoreIfNull]
